U5 t44 capacitor unit entity

See merge request capacitor-archive/capacitor-archive!59

+++ server/src/main/java/com/example/demo/repository/ConstructionRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Construction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConstructionRepository extends JpaRepository<Construction, Integer> {
+public interface ConstructionRepository extends JpaRepository<Construction, Long> {
 }
